Moved KML writer to utils
